requirements.txt
azure-functions
azure-identity
azure-keyvault-secrets
sshkey-tools
